Edit index.html and style.css
These edits were made in an effort to re-evaluate the method I chose to
approach this project. Initially I had tried to take care of ALL the
html and then doing the css, but I ran into a lot of margin and padding
issues. Now, I'm taking the project one section at a time, doing the
html first and styling it with css. This is the first completed section
of the project.

The dimensions don't need to be perfect, which is the only saving grace
I have for completing this project. Overall, it should be pretty alike.
I'll be working on the next section as soon as possible.